Fixes unaligned memory access on non-x86 or x86_64 platforms. Closes GH-1646.
The SIGBUS on Solaris SPARC was an alignment problem. It was caused by casting
pointers-to-arrays-of-bytes into pointers-to-arrays-of-4-byte-ints, and then
dereferencing them even when they weren't aligned (`addr % 4 != 0`). So I
replaced the dereferencing of those pointers with 4 individual loads each.
However after fixing that I found that the actual decoding was broken. Decoding
a base64 string gave be back the string
"fulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfulfu"
instead of what it was supposed to be. After investigating, it was revealed that
the compiler was doing surprising optimizations on the code based on the fact that
casting between 1-byte ints and 4-byte ints the way that this code did violates
strict aliasing rules and counts as UB. So I rewrote the function in standards
compliant C. Then I simply flagged the strict aliasing violating version as
x86/x86_64 only, and used the standards compliant version elsewhere.
